chore: move miner to separate module
